+++ src/semantic_release/cli/commands/publish.py
@@ -8,7 +8,7 @@ from git import Repo
 
 from semantic_release.cli.util import noop_report
 from semantic_release.hvcs.remote_hvcs_base import RemoteHvcsBase
-from semantic_release.version import tags_and_versions
+from semantic_release.version.algorithm import tags_and_versions
 
 if TYPE_CHECKING:
     from semantic_release.cli.cli_context import CliContextObj
